Project/MSTR-API/pyproject.toml
[project]
name = "mstr-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "easycharts>=0.106",
    "fastapi>=0.115.5",
    "pandas>=2.2.3",
    "pandas-datareader>=0.10.0",
    "polygon-api-client>=1.14.2",
    "requests>=2.32.3",
    "uvicorn>=0.32.1",
    "yfinance>=0.2.50",
]
